split init script
